fix: address review findings — break circular import, use shared types
Move SearchPriority and IssueCandidate to types.ts to break the
circular import between issue-discovery.ts and issue-vetting.ts.
Make isRateLimitError proxy public (tests access it). Use
ViabilityScoreParams type instead of redeclaring inline params.
